requirements.txt
pefile~=2023.2.7
pyelftools~=0.29
pythonnet~=3.0.1